fix: next router skipping signed-in check
currently when a user signs into OWA and subsequently clicks the "Sign in"
link in Aila they're presented with the error

"You're currently in single session mode. You can only be signed into one account at a time."

This appears to be caused by navigating to the page through the Next router. Replacing it with
a plain anchor appears to trigger the expected signed-in check and redirect the user back to the app.